Release 1.4.1 compiler warning correction

+++ src/LSS.h
@@ -240,7 +240,7 @@ class LSS
 {
 public:
 	// Public functions - Class
-	static void setReadTimeouts(long start_response_timeout=LSS_Timeout, long msg_char_timeout=LSS_Timeout);
+	static void setReadTimeouts(uint32_t start_response_timeout=LSS_Timeout, uint32_t msg_char_timeout=LSS_Timeout);
 	static int timedRead(void);
 	bool charToInt(char * inputstr, int32_t * intnum);
 	//static void initBus(Stream &, uint32_t);
@@ -346,7 +346,7 @@ private:
 	static LSS_LastCommStatus lastCommStatus;
 	static volatile unsigned int readID;
 	static char value[24];
-	static long _msg_char_timeout;   // timeout waiting for characters inside of packet
+	static uint32_t _msg_char_timeout;   // timeout waiting for characters inside of packet
 	// Private functions - Instance
 
 	// Private attributes - Instance
